Refactor CPE AL
Remove version specifications.
Split boolean expression tests into generic and platform-related.
Fix for mutable default argument value.
